previous use of log10 wouldn't compile. fixed
git-svn-id: http://code.icecube.wisc.edu/svn/projects/phys-services/trunk@12405 16731396-06f5-0310-8873-f7f720988828

+++ private/phys-source/I3CalibrateStatusModule.cxx
@@ -65,7 +65,7 @@ DoTheCalibration(I3RawDOMStatusPtr rawstatus,
 
   double predictedSpeMean;
   double log_gain = 0.0;
-  int currentVoltage=(unsigned int)(rawstatus->GetPMTHV()/I3Units::volt);
+  double currentVoltage=(rawstatus->GetPMTHV()/I3Units::volt);
   const LinearFit hvgain = calib->GetHVGainFit();
   
   if(currentVoltage >0.0)
